minor validation changes on auth controller

+++ app/Http/Controllers/AuthController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use App\Models\User;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 
 class AuthController extends Controller
@@ -13,15 +12,15 @@ class AuthController extends Controller
     public function login(Request $request): JsonResponse
     {
         $request->validate([
-            'username' => 'required',
-            'password' => 'required'
+            'username' => 'required|string|min:6',
+            'password' => 'required|string|min:6'
         ]);
 
         $user = User::where('username', $request->username)->first();
 
         if (!$user || !Hash::check($request->password, $user->password)) {
             return response()->json([
-                'message' => 'Invalid credentials',
+                'message' => 'Invalid username or password',
                 'success' => false
             ], 401);
         }
@@ -40,23 +39,32 @@ class AuthController extends Controller
     {   
         $user = User::where('username', $request->username)->first();
 
+        if (!$user) {
+            return response()->json([
+                'message' => 'User not found',
+                'success' => false
+            ], 404); 
+        }
+
         if ($user->tokens()->delete()) {
             return response()->json([
                 'message' => 'Logout succesful',
+                'success' => true
             ]);
         }
 
         return response()->json([
-            'message' => 'Not found'
-        ], 404); 
+            'message' => "Server error",
+            'success' => false
+        ], 500);
     }
 
     public function register(Request $request): JsonResponse
     {   
         $request->validate([
-            'username' => 'required',
-            'email' => 'required',
-            'password' => 'required'
+            'username' => 'required|string|min:6',
+            'email' => 'required|string|email',
+            'password' => 'required|string|min:6'
         ]);
 
         $user = User::create([
@@ -92,12 +100,14 @@ class AuthController extends Controller
         if ($user) {
             return response()->json([
                 'message' => 'User retrieved',
-                'data' => $user
+                'data' => $user,
+                'success' => true
             ]);
         }
 
         return response()->json([
             'message' => 'Not authenticated',
+            'success' => false
         ], 401); 
     }
 }
